+++ src/backend/commands/admin-command.manager.ts
@@ -1,0 +1,18 @@
+import {SiteMapGeneratorCommand} from './sitemap-generator.command';
+import {
+    CommonAdminCommandConfigType,
+    CommonAdminCommandManager
+} from '@dps/mycms-server-commons/dist/backend-commons/commands/common-admin-command.manager';
+
+export interface AdminCommandConfigType extends CommonAdminCommandConfigType {
+}
+
+export class AdminCommandManager extends CommonAdminCommandManager<AdminCommandConfigType> {
+    constructor(adminCommandConfig: AdminCommandConfigType) {
+        super({
+            'generateSitemap': new SiteMapGeneratorCommand()
+        }, adminCommandConfig);
+    }
+
+}
+