+++ src/backend/server.ts
@@ -7,10 +7,10 @@ import {ServerConfig, ServerModuleLoader} from './server-module.loader';
 const argv = minimist(process.argv.slice(2));
 
 const debug = argv['debug'] || false;
-const filePathConfigJson = argv['c'] || argv['backend'];
-const filePathFirewallConfigJson = argv['f'] || argv['firewall'];
+const filePathConfigJson = argv['backend'];
+const filePathFirewallConfigJson = argv['firewall'];
 if (filePathConfigJson === undefined || filePathFirewallConfigJson === undefined) {
-    console.error('ERROR - parameters required backendConfig: "-c | --backend" firewallConfig: "-f | --firewall"');
+    console.error('ERROR - parameters required backendConfig: "--backend" firewallConfig: "--firewall"');
     process.exit(-1);
 }
 
@@ -29,9 +29,11 @@ const app = express();
 // load modules
 ServerModuleLoader.loadModules(app, serverConfig);
 
-// start server
-app.listen(serverConfig.backendConfig.port, function () {
-    console.log('MyStarTrek app listening on port ' + serverConfig.backendConfig['port']);
+// start server as seen on https://nodejs.org/api/net.html#net_server_listen
+const bindIp = serverConfig.backendConfig.bindIp ? serverConfig.backendConfig.bindIp : '127.0.0.1';
+const tcpBacklog = serverConfig.backendConfig.tcpBacklog ? serverConfig.backendConfig.tcpBacklog : 511;
+app.listen(serverConfig.backendConfig.port, bindIp,  tcpBacklog, function () {
+    console.log('MyStarTrek app listening on ip/port/tcpBacklog', bindIp, serverConfig.backendConfig.port, tcpBacklog);
     if (!debug) {
         console.log = function() {};
     }