+++ src/backend/serverAdmin.ts
@@ -1,11 +1,22 @@
 import minimist from 'minimist';
-import {SiteMapGeneratorCommand} from './commands/sitemap-generator.command';
-import {utils} from 'js-data';
-import {StarDocLoaderCommand} from "./commands/sdoc-loader.command";
-import {StarDocExporterCommand} from "./commands/sdoc-exporter.command";
-import {StarDocConverterCommand} from "./commands/sdoc-converter.command";
+import {AdminCommandConfigType, AdminCommandManager} from './commands/admin-command.manager';
+import fs from 'fs';
 
-const argv = minimist(process.argv.slice(2));
+const argv: string[] = minimist(process.argv.slice(2));
+
+const filePathConfigJson = argv['adminclibackend'];
+if (filePathConfigJson === undefined) {
+    console.error('ERROR - parameters required adminclibackend: "--adminclibackend"');
+    process.exit(-1);
+}
+
+const adminBackendConfig: AdminCommandConfigType = JSON.parse(fs.readFileSync(filePathConfigJson, { encoding: 'utf8' }));
+const adminCommandManager = new AdminCommandManager(adminBackendConfig);
+if (argv['help'] || argv['usage']) {
+    console.error('help -> prepared commands: ', JSON.stringify(adminCommandManager.listPreparedCommands(), null, 2));
+    console.error('help -> available commands and parameters: ', JSON.stringify(adminCommandManager.listAvailableCommands(), null, 2));
+    process.exit(0);
+}
 
 // disable debug-logging
 const debug = argv['debug'] || false;
@@ -17,31 +28,7 @@ if (!debug || debug === false || parseInt(debug, 10) < 1) {
     console.debug = function() {};
 }
 
-const sdocConverter = new StarDocConverterCommand();
-const sdocLoader = new StarDocLoaderCommand();
-const sdocExporter = new StarDocExporterCommand();
-const siteMapGenerator = new SiteMapGeneratorCommand();
-
-let promise: Promise<any>;
-switch (argv['command']) {
-    case 'generateSitemap':
-        promise = siteMapGenerator.process(argv);
-        break;
-    case 'convertStarDoc':
-        promise = sdocConverter.process(argv);
-        break;
-    case 'loadStarDoc':
-        promise = sdocLoader.process(argv);
-        break;
-    case 'exportStarDoc':
-        promise = sdocExporter.process(argv);
-        break;
-    default:
-        console.error('unknown command:', argv);
-        promise = utils.reject('unknown command');
-}
-
-promise.then(value => {
+adminCommandManager.runCommand(argv).then(value => {
     console.log('DONE - command finished:', value, argv);
     process.exit(0);
 }).catch(reason => {