+++ src/frontendserver/frontendserverAngular.ts
@@ -21,10 +21,10 @@ const staticFolder = join(process.cwd(), 'dist/static');
 const distProfile = 'DIST_PROFILE';
 const distServerProfile = 'DIST_SERVER_PROFILE';
 const filePathConfigJson = argv['frontend'];
-const filePathBackendConfigJson = argv['c'] || argv['backend'];
-const filePathFirewallConfigJson = argv['f'] || argv['firewall'];
+const filePathBackendConfigJson = argv['backend'];
+const filePathFirewallConfigJson = argv['firewall'];
 if (filePathConfigJson === undefined || filePathBackendConfigJson === undefined || filePathFirewallConfigJson === undefined) {
-    console.error('ERROR - parameters required frontendConfig:  "--frontend" backendConfig: "-c | --backend" firewallConfig: "-f | --firewall"');
+    console.error('ERROR - parameters required frontendConfig:  "--frontend" backendConfig: "--backend" firewallConfig: "--firewall"');
     process.exit(-1);
 }
 
@@ -33,7 +33,9 @@ export interface ServerConfig {
     backendConfig: BackendConfigType;
     firewallConfig: FirewallConfig;
     frontendConfig: {
+        bindIp: string,
         port: number,
+        tcpBacklog: number,
         cacheFolder: string,
         redirectFileJson?: string,
         redirectOnlyCached?: boolean
@@ -67,8 +69,10 @@ AngularUniversalFrontendServerModule.configureDefaultServer(app, frontendConfig)
 ConfigureServerModule.configureDefaultErrorHandler(app);
 
 // Start up the Node server
-app.listen(serverConfig.frontendConfig.port, function () {
-    console.log('MyStarTrek app listening on port ' + serverConfig.frontendConfig.port);
+const bindIp = serverConfig.frontendConfig.bindIp ? serverConfig.frontendConfig.bindIp : '127.0.0.1';
+const tcpBacklog = serverConfig.frontendConfig.tcpBacklog ? serverConfig.frontendConfig.tcpBacklog : 511;
+app.listen(serverConfig.frontendConfig.port, bindIp,  tcpBacklog, function () {
+    console.log('MyStarTrek app listening on ip/port/tcpBacklog', bindIp, serverConfig.frontendConfig.port, tcpBacklog);
 
     // disable debug-logging
     if (!debug) {