+++ docs/CHANGELOG.md
@@ -2,6 +2,30 @@
  
 # Versions
  
+## 4.0.0
+- improved security
+- added admin-server and parameter-validation on serverAdmin
+- use commons-5.2.0
+
+### new features
+- backend: added admin-server
+- frontend: added admin-area
+
+### improvements
+- security: bind on localhost only by default
+- backend: added parameter-validation on serverAdmin
+- backend: added scripts to prepare app-environment on startup
+- install: for derived apps - optional use local solr 
+- build: use commons-5.2.0
+
+### bug fixes
+- none
+ 
+### breaking changes
+- backend: removed short-parameters for configfiles -c -> --backend, -f -> --firewall or --file, -s -> --sitemap 
+- backend: serverAdmin requires action-parameters
+
+
 ## 3.0.0
 - now bases on mysimplehomepage-3.0.0
 - improved build-process